chore: dedup & merge composeAutoExternalConfig to config.ts
