+++ packages/core/src/config.ts
@@ -11,6 +11,7 @@ import {
 import glob from 'fast-glob';
 import { DEFAULT_CONFIG_NAME, DEFAULT_EXTENSIONS } from './constant';
 import type {
+  AutoExternal,
   Format,
   LibConfig,
   PkgJson,
@@ -21,7 +22,6 @@ import type {
   Syntax,
 } from './types';
 import { getDefaultExtension } from './utils/extension';
-import { composeAutoExternalConfig } from './utils/external';
 import {
   calcLongestCommonPath,
   color,
@@ -81,6 +81,60 @@ export async function loadConfig(
 
   return content as RslibConfig;
 }
+
+export const composeAutoExternalConfig = (options: {
+  autoExternal: AutoExternal;
+  pkgJson?: PkgJson;
+  userExternals?: NonNullable<RsbuildConfig['output']>['externals'];
+}): RsbuildConfig => {
+  const { autoExternal, pkgJson, userExternals } = options;
+
+  if (!autoExternal) {
+    return {};
+  }
+
+  if (!pkgJson) {
+    logger.warn(
+      'autoExternal configuration will not be applied due to read package.json failed',
+    );
+    return {};
+  }
+
+  const externalOptions = {
+    dependencies: true,
+    peerDependencies: true,
+    devDependencies: false,
+    ...(autoExternal === true ? {} : autoExternal),
+  };
+
+  // User externals configuration has higher priority than autoExternal
+  // eg: autoExternal: ['react'], user: output: { externals: { react: 'react-1' } }
+  // Only handle the case where the externals type is object, string / string[] does not need to be processed, other types are too complex.
+  const userExternalKeys =
+    userExternals &&
+    Object.prototype.toString.call(userExternals) === '[object Object]'
+      ? Object.keys(userExternals)
+      : [];
+
+  const externals = (
+    ['dependencies', 'peerDependencies', 'devDependencies'] as const
+  )
+    .reduce<string[]>((prev, type) => {
+      if (externalOptions[type]) {
+        return pkgJson[type] ? prev.concat(Object.keys(pkgJson[type]!)) : prev;
+      }
+      return prev;
+    }, [])
+    .filter((name) => !userExternalKeys.includes(name));
+
+  return externals.length
+    ? {
+        output: {
+          externals: Array.from(new Set(externals)),
+        },
+      }
+    : {};
+};
 
 export async function createInternalRsbuildConfig(): Promise<RsbuildConfig> {
   return defineRsbuildConfig({
